refactor: let staged own review comment actions
Replace the diff-viewer package Note/Commit/GitHub-specific comment editor props with a generic host-rendered commentActions snippet. Move the Staged review comment buttons, state types, icons, styling, and click routing into a Staged-owned ReviewCommentActions component while preserving the existing inline comment behavior.
