+++ packages/diff-viewer/src/lib/components/CommentEditor.svelte
@@ -5,17 +5,9 @@
   Handles its own visibility based on scroll position.
 -->
 <script lang="ts">
-  import { Check, FileText, GitCommitVertical, Github, Loader2, Trash2 } from 'lucide-svelte';
-  import type { Comment } from '../types';
-
-  export type GithubButtonState = 'idle' | 'sending' | 'sent' | 'stale';
-
-  /**
-   * State of a comment's "Note"/"Commit" button, derived from the linked
-   * session's status by the parent. `queued` is treated as `running`, and
-   * `error`/`cancelled` collapse back to `idle` so the user can retry.
-   */
-  export type CommentSessionState = 'idle' | 'running' | 'completed';
+  import { Trash2 } from 'lucide-svelte';
+  import type { Snippet } from 'svelte';
+  import type { Comment, CommentActionContext } from '../types';
 
   interface Props {
     /** Position relative to the viewer container */
@@ -36,18 +28,8 @@
     onCancel: () => void;
     /** Called when comment is deleted (only shown if existingComment is set) */
     onDelete?: () => void;
-    /** Called when "Note" action is clicked (only shown if existingComment is set). */
-    onNote?: (event: MouseEvent) => void;
-    /** Called when "Commit" action is clicked (only shown if existingComment is set). */
-    onCommit?: (event: MouseEvent) => void;
-    /** Called when "GitHub" action is clicked (only shown if existingComment is set). */
-    onGithub?: () => void;
-    /** Current state of the GitHub send/update button. */
-    githubState?: GithubButtonState;
-    /** Current state of the "Note" button, reflecting its linked session. */
-    noteState?: CommentSessionState;
-    /** Current state of the "Commit" button, reflecting its linked session. */
-    commitState?: CommentSessionState;
+    /** Host-rendered actions for existing comments. */
+    commentActions?: Snippet<[CommentActionContext]>;
   }
 
   let {
@@ -61,12 +43,7 @@
     onSubmit,
     onCancel,
     onDelete,
-    onNote,
-    onCommit,
-    onGithub,
-    githubState = 'idle',
-    noteState = 'idle',
-    commitState = 'idle',
+    commentActions,
   }: Props = $props();
 
   // Track current input value - initialized by effect when existingComment changes
@@ -138,73 +115,9 @@
     <span class="comment-editor-help">
       {readOnly ? 'Read-only · Esc to close' : 'Enter to save · Esc to cancel'}
     </span>
-    {#if existingComment && (onNote || onCommit || onGithub)}
+    {#if existingComment && commentActions}
       <div class="comment-action-buttons">
-        {#if onNote}
-          <button
-            class="comment-action-btn note-btn"
-            class:session-active={noteState !== 'idle'}
-            onclick={(e) => onNote?.(e)}
-            title={noteState === 'running'
-              ? 'Note session in progress'
-              : noteState === 'completed'
-                ? 'Open note'
-                : 'New note (Option+click to skip dialog)'}
-          >
-            {#if noteState === 'running'}
-              <Loader2 size={12} class="spinner" />
-            {:else}
-              <FileText size={12} />
-            {/if}
-            <span>Note</span>
-          </button>
-        {/if}
-        {#if onCommit}
-          <button
-            class="comment-action-btn commit-btn"
-            class:session-active={commitState !== 'idle'}
-            onclick={(e) => onCommit?.(e)}
-            title={commitState === 'running'
-              ? 'Commit session in progress'
-              : commitState === 'completed'
-                ? 'Show commit'
-                : 'New commit (Option+click to skip dialog)'}
-          >
-            {#if commitState === 'running'}
-              <Loader2 size={12} class="spinner" />
-            {:else}
-              <GitCommitVertical size={12} />
-            {/if}
-            <span>Commit</span>
-          </button>
-        {/if}
-        {#if onGithub}
-          <button
-            class="comment-action-btn github-btn"
-            class:github-btn-sent={githubState === 'sent'}
-            onclick={() => onGithub?.()}
-            title={githubState === 'sent'
-              ? 'Open GitHub comment'
-              : githubState === 'stale'
-                ? 'Update on GitHub'
-                : 'Send to GitHub'}
-            disabled={githubState === 'sending'}
-          >
-            {#if githubState === 'sending'}
-              <Loader2 size={12} class="spinner" />
-            {:else if githubState === 'sent'}
-              <Check size={12} class="github-sent-check" />
-              <Github size={12} />
-            {:else}
-              <Github size={12} />
-            {/if}
-            {#if githubState === 'stale'}
-              <span>Update on GitHub</span>
-            {:else}
-              <span>GitHub</span>
-            {/if}
-          </button>
-        {/if}
+        {@render commentActions({ comment: existingComment })}
       </div>
     {/if}
     {#if existingComment && onDelete}
@@ -276,105 +189,6 @@
     display: flex;
     align-items: center;
     gap: 4px;
-  }
-
-  .comment-action-btn {
-    display: inline-flex;
-    align-items: center;
-    gap: 4px;
-    padding: 2px 6px;
-    border-radius: 4px;
-    border: 1px dashed var(--border-subtle);
-    background: transparent;
-    color: var(--text-muted);
-    font-size: calc(var(--size-xs) - 1px);
-    font-weight: 500;
-    cursor: pointer;
-    transition:
-      color 0.15s,
-      border-color 0.15s,
-      background-color 0.15s;
-  }
-
-  .comment-action-btn.note-btn :global(svg) {
-    color: var(--note-color);
-  }
-
-  .comment-action-btn.commit-btn :global(svg) {
-    color: var(--commit-color);
-  }
-
-  .comment-action-btn.github-btn :global(svg) {
-    color: var(--text-primary);
-  }
-
-  .comment-action-btn.note-btn:hover {
-    color: var(--note-color);
-    border-color: var(--note-color);
-    background-color: var(--note-bg);
-  }
-
-  .comment-action-btn.commit-btn:hover {
-    color: var(--commit-color);
-    border-color: var(--commit-color);
-    background-color: var(--commit-bg);
-  }
-
-  /* Running/completed state — mirrors the styled spinner-with-background and
-     completed icon used on Branch Card timeline rows (TimelineRow.svelte). */
-  .comment-action-btn.note-btn.session-active {
-    color: var(--note-color);
-    border-color: transparent;
-    background-color: var(--note-bg);
-  }
-
-  .comment-action-btn.commit-btn.session-active {
-    color: var(--commit-color);
-    border-color: transparent;
-    background-color: var(--commit-bg);
-  }
-
-  .comment-action-btn.github-btn:hover:not(:disabled) {
-    color: var(--text-primary);
-    border-color: var(--text-muted);
-    background-color: var(--bg-hover);
-  }
-
-  .comment-action-btn.github-btn:disabled {
-    cursor: default;
-    opacity: 0.7;
-  }
-
-  .comment-action-btn.github-btn-sent {
-    border-style: solid;
-    color: var(--text-primary);
-    border-color: var(--status-added, #3fb950);
-  }
-
-  .comment-action-btn.github-btn-sent :global(.github-sent-check) {
-    color: var(--status-added, #3fb950) !important;
-  }
-
-  /* Spin in place like the app's shared Spinner. Without an explicit
-     transform-box/origin, WebKit rotates the SVG about an off-center point,
-     making the icon visibly wobble; pin the origin to the exact viewBox
-     center (and promote to a layer) so it spins cleanly. */
-  .comment-action-btn :global(.spinner) {
-    display: block;
-    overflow: visible;
-    transform-box: view-box;
-    transform-origin: 50% 50%;
-    backface-visibility: hidden;
-    animation: spin 1s linear infinite;
-  }
-
-  @keyframes spin {
-    from {
-      transform: rotate(0deg);
-    }
-    to {
-      transform: rotate(360deg);
-    }
   }
 
   .delete-comment-btn {
