+++ packages/diff-viewer/src/lib/components/index.ts
@@ -1,7 +1,7 @@
 export { default as DiffViewer } from './DiffViewer.svelte';
 export { default as ImageDiffViewer } from './ImageDiffViewer.svelte';
 export { default as CommentEditor } from './CommentEditor.svelte';
-export type { CommentSessionState } from './CommentEditor.svelte';
+export type { CommentActionContext } from '../types';
 export { default as AnnotationOverlay } from './AnnotationOverlay.svelte';
 export { default as BeforeAnnotationOverlay } from './BeforeAnnotationOverlay.svelte';
 export { default as Scrollbar } from './Scrollbar.svelte';
